change entrance of program + do README.md

+++ clear.sh
@@ -1,2 +1,6 @@
 #!/bin/bash
-git clean -dfX
+if [[ -f ".gitignore" ]];then
+  git clean -dfX
+else
+  echo "no .gitignore file found. clean does not success."
+fi
